+++ app/src/main/java/com/theunquenchedservant/granthornersbiblereadingsystem/service/AlarmReceiver.kt
@@ -1,4 +1,4 @@
-package com.theunquenchedservant.granthornersbiblereadingsystem.ui.notifications
+package com.theunquenchedservant.granthornersbiblereadingsystem.service
 
 import android.app.Notification
 import android.app.NotificationManager
