+++ app/src/main/java/com/theunquenchedservant/granthornersbiblereadingsystem/ui/home/HomeFragment.kt
@@ -30,13 +30,13 @@ import com.theunquenchedservant.granthornersbiblereadingsystem.R
 import com.theunquenchedservant.granthornersbiblereadingsystem.Marker.markAll
 import com.theunquenchedservant.granthornersbiblereadingsystem.Marker.markSingle
 import com.theunquenchedservant.granthornersbiblereadingsystem.SharedPref
-import com.theunquenchedservant.granthornersbiblereadingsystem.ui.notifications.DailyCheck
+import com.theunquenchedservant.granthornersbiblereadingsystem.utilities.DailyCheck
 import com.theunquenchedservant.granthornersbiblereadingsystem.MainActivity
 import com.theunquenchedservant.granthornersbiblereadingsystem.databinding.ActivityMainBinding
 import com.theunquenchedservant.granthornersbiblereadingsystem.databinding.FragmentHomeBinding
-import com.theunquenchedservant.granthornersbiblereadingsystem.ui.notifications.AlarmCreator.createAlarm
-import com.theunquenchedservant.granthornersbiblereadingsystem.ui.notifications.AlarmCreator.createAlarms
-import com.theunquenchedservant.granthornersbiblereadingsystem.ui.notifications.AlarmCreator.createNotificationChannel
+import com.theunquenchedservant.granthornersbiblereadingsystem.service.AlarmCreator.createAlarm
+import com.theunquenchedservant.granthornersbiblereadingsystem.service.AlarmCreator.createAlarms
+import com.theunquenchedservant.granthornersbiblereadingsystem.service.AlarmCreator.createNotificationChannel
 import com.theunquenchedservant.granthornersbiblereadingsystem.utilities.SharedPref.getBoolPref
 import com.theunquenchedservant.granthornersbiblereadingsystem.utilities.SharedPref.getIntPref
 import com.theunquenchedservant.granthornersbiblereadingsystem.utilities.SharedPref.setIntPref
