+++ app/src/main/java/com/theunquenchedservant/granthornersbiblereadingsystem/ui/settings/NotificationFragment.kt
@@ -5,8 +5,8 @@ import androidx.fragment.app.DialogFragment
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
 import com.theunquenchedservant.granthornersbiblereadingsystem.R
-import com.theunquenchedservant.granthornersbiblereadingsystem.ui.notifications.AlarmCreator.cancelAlarm
-import com.theunquenchedservant.granthornersbiblereadingsystem.ui.notifications.AlarmCreator.createAlarm
+import com.theunquenchedservant.granthornersbiblereadingsystem.service.AlarmCreator.cancelAlarm
+import com.theunquenchedservant.granthornersbiblereadingsystem.service.AlarmCreator.createAlarm
 import com.theunquenchedservant.granthornersbiblereadingsystem.utilities.SharedPref.getIntPref
 import com.theunquenchedservant.granthornersbiblereadingsystem.utilities.SharedPref.setBoolPref
 import com.theunquenchedservant.granthornersbiblereadingsystem.utilities.SharedPref.setIntPref
